+++ schemacode/src/main/java/info/teksol/schemacode/schema/BlockOrder.java
@@ -1,0 +1,7 @@
+package info.teksol.schemacode.schema;
+
+public enum BlockOrder {
+    ORIGINAL,
+    HORIZONTAL,
+    VERTICAL
+}